+++ modules/frontend/tracebyidsharding.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
-	"github.com/golang/protobuf/proto"  //nolint:all //deprecated 
+	"github.com/golang/protobuf/proto" //nolint:all //deprecated
 	"github.com/grafana/tempo/modules/querier"
 	"github.com/grafana/tempo/pkg/api"
 	"github.com/grafana/tempo/pkg/model/trace"